+++ extensions/logtail/src/hooks/useDefaultSourceId.ts
@@ -1,6 +1,6 @@
 import { LocalStorage } from "@raycast/api";
 import { useCachedPromise } from "@raycast/utils";
-import { LogTail } from "../lib/logtail";
+import { Logtail } from "../lib/logtail";
 import { UseCachedPromiseReturnType } from "@raycast/utils/dist/types";
 
 export type UseDefaultSourceIdResult = UseCachedPromiseReturnType<string, undefined> & {
@@ -9,15 +9,17 @@ export type UseDefaultSourceIdResult = UseCachedPromiseReturnType<string, undefi
 };
 
 export const useDefaultSourceId = (): UseDefaultSourceIdResult => {
-  const result = useCachedPromise(() => LocalStorage.getItem<string>(LogTail.DEFAULT_SOURCE_ID_CACHE_KEY));
+  const result = useCachedPromise(() => LocalStorage.getItem<string>(Logtail.DEFAULT_SOURCE_ID_CACHE_KEY));
 
   return {
     ...result,
     removeDefaultSourceId: async () => {
-      await LocalStorage.removeItem(LogTail.DEFAULT_SOURCE_ID_CACHE_KEY);
+      await LocalStorage.removeItem(Logtail.DEFAULT_SOURCE_ID_CACHE_KEY);
+      await result.mutate(undefined, { optimisticUpdate: () => undefined });
     },
     setDefaultSourceId: async (sourceId: string) => {
-      await LocalStorage.setItem(LogTail.DEFAULT_SOURCE_ID_CACHE_KEY, sourceId);
+      await LocalStorage.setItem(Logtail.DEFAULT_SOURCE_ID_CACHE_KEY, sourceId);
+      await result.mutate(undefined, { optimisticUpdate: () => sourceId });
     },
   };
 };