+++ extensions/logtail/CHANGELOG.md
@@ -1,12 +1,15 @@
-# LogTail Changelog
+# Logtail Changelog
 
-## [Initial Version] - 2023-05-11
+## [Initial Version] - 2023-05-23
 
 - Added Query Logs Command
   - Allow saving query for future use
   - Allow copying host and message fields
   - Allow copying raw log body
   - Allow querying for other occurrences of the selected message
+  - Allow viewing sources
+  - Allow automatically using a source for query if only one source is found
+  - Allow setting/removing the default source to use for future queries
 - Added View Saved Queries Command
   - Allow removing saved queries
 - Added Set Default Source to query logs from
@@ -16,4 +19,3 @@
 - Added View Metadata Tags Command
   - Allow removing a metadata tag
   - Allow editing an existing metadata tag
-- Added Configure Command for authenticating with LogTail